requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-socketio==5.4.0
python-engineio==4.2.1
SpeechRecognition==3.8.1
python-dotenv==0.19.0
